requirements.txt
python-telegram-bot==20.2
sarufi==0.1.2
python-dotenv==1.0.0
